ref(spans): restructure GCP log analyzer into log_analyzer package
- Add src/sentry/spans/log_analyzer/ with fetchers.py and analyzer.py
- fetchers: TraceSpans, LogEntry, parse_top_traces, build_gcp_filter,
  fetch_logs_from_file, fetch_logs_from_gcp
- analyzer: TraceKey, TraceStats, SummaryStats, LogAnalyzer (with
  _trace_stats cache)
- Keep gcp_log_analyzer.py as CLI and formatting only (imports from
  log_analyzer)
- Add tests/sentry/spans/log_analyzer/ with test_analyzer.py and
  test_fetchers.py; convert tests to top-level functions; parametrize
  build_gcp_filter tests
- Remove unused typing.Any import from fetchers.py (flake8)
