+++ src/sentry/spans/log_analyzer/__init__.py
@@ -1,0 +1,23 @@
+"""Log analyzer for span buffer slow EVALSHA operations."""
+
+from sentry.spans.log_analyzer.analyzer import LogAnalyzer, SummaryStats, TraceStats
+from sentry.spans.log_analyzer.fetchers import (
+    LogEntry,
+    TraceSpans,
+    build_gcp_filter,
+    fetch_logs_from_file,
+    fetch_logs_from_gcp,
+    parse_top_traces,
+)
+
+__all__ = [
+    "LogAnalyzer",
+    "LogEntry",
+    "SummaryStats",
+    "TraceSpans",
+    "TraceStats",
+    "build_gcp_filter",
+    "fetch_logs_from_file",
+    "fetch_logs_from_gcp",
+    "parse_top_traces",
+]
