+++ tests/sentry/spans/log_analyzer/__init__.py
@@ -1,0 +1,1 @@
+# Tests for sentry.spans.log_analyzer
